made the description better
We need a hero img.  I am thinking that a picture of NHS would be good.  I have an exact idea actually for the picture, but it is pretty hard to explain.  If there is a photography team or the yearbook class or something, we could get them to take the picture.